:hammer: REFACTOR (project): extract project preview from page
